Moved mock api classes into the tests/api folder
This is to start the removal of the pytest code from the app code
