+++ api/admin/controller/search_service_self_tests.py
@@ -3,10 +3,9 @@ from flask_babel import lazy_gettext as _
 from api.admin.controller.self_tests import SelfTestsController
 from core.external_search import ExternalSearchIndex
 from core.model import ExternalIntegration
-from core.testing import ExternalSearchTest
 
 
-class SearchServiceSelfTestsController(SelfTestsController, ExternalSearchTest):
+class SearchServiceSelfTestsController(SelfTestsController):
     def __init__(self, manager):
         super().__init__(manager)
         self.type = _("search service")
