+++ api/odl.py
@@ -42,7 +42,6 @@ from core.model import (
     RightsStatus,
     Session,
     get_one,
-    get_one_or_create,
 )
 from core.model.configuration import (
     ConfigurationAttributeType,
@@ -57,7 +56,6 @@ from core.model.licensing import LicenseStatus
 from core.model.patron import Patron
 from core.monitor import CollectionMonitor
 from core.opds_import import OPDSImporter, OPDSImportMonitor, OPDSXMLParser
-from core.testing import DatabaseTest, MockRequestsResponse
 from core.util.datetime_helpers import to_utc, utc_now
 from core.util.http import HTTP, BadResponseException, RemoteIntegrationException
 from core.util.string_helpers import base64
@@ -1877,41 +1875,3 @@ class SharedODLImportMonitor(OPDSImportMonitor):
     def opds_url(self, collection):
         base_url = collection.external_account_id
         return base_url + "/crawlable"
-
-
-class MockSharedODLAPI(SharedODLAPI):
-    """Mock API for tests that overrides _get and tracks requests."""
-
-    @classmethod
-    def mock_collection(cls, _db):
-        """Create a mock ODL collection to use in tests."""
-        library = DatabaseTest.make_default_library(_db)
-        collection, ignore = get_one_or_create(
-            _db,
-            Collection,
-            name="Test Shared ODL Collection",
-            create_method_kwargs=dict(
-                external_account_id="http://shared-odl",
-            ),
-        )
-        integration = collection.create_external_integration(protocol=SharedODLAPI.NAME)
-        library.collections.append(collection)
-        return collection
-
-    def __init__(self, _db, collection, *args, **kwargs):
-        self.responses = []
-        self.requests = []
-        self.request_args = []
-        super().__init__(_db, collection, *args, **kwargs)
-
-    def queue_response(self, status_code, headers={}, content=None):
-        self.responses.insert(0, MockRequestsResponse(status_code, headers, content))
-
-    def _get(self, url, patron=None, headers=None, allowed_response_codes=None):
-        allowed_response_codes = allowed_response_codes or ["2xx", "3xx"]
-        self.requests.append(url)
-        self.request_args.append((patron, headers, allowed_response_codes))
-        response = self.responses.pop()
-        return HTTP._process_response(
-            url, response, allowed_response_codes=allowed_response_codes
-        )
