+++ core/testing.py
@@ -1114,75 +1114,6 @@ class SearchClientForTesting(ExternalSearchIndex):
         return super().setup_index(new_index, number_of_shards=1, number_of_replicas=0)
 
 
-@pytest.mark.opensearch
-class ExternalSearchTest(DatabaseTest):
-    """
-    These tests require opensearch to be running locally. If it's not, or there's
-    an error creating the index, the tests will pass without doing anything.
-
-    Tests for opensearch are useful for ensuring that we haven't accidentally broken
-    a type of search by changing analyzers or queries, but search needs to be tested manually
-    to ensure that it works well overall, with a realistic index.
-    """
-
-    SIMPLIFIED_TEST_OPENSEARCH = os.environ.get(
-        "SIMPLIFIED_TEST_OPENSEARCH", "http://localhost:9200"
-    )
-
-    def setup_method(self):
-
-        super().setup_method()
-
-        # Track the indexes created so they can be torn down at the
-        # end of the test.
-        self.indexes = []
-
-        self.integration = self._external_integration(
-            ExternalIntegration.OPENSEARCH,
-            goal=ExternalIntegration.SEARCH_GOAL,
-            url=self.SIMPLIFIED_TEST_OPENSEARCH,
-            settings={
-                ExternalSearchIndex.WORKS_INDEX_PREFIX_KEY: "test_index",
-                ExternalSearchIndex.TEST_SEARCH_TERM_KEY: "test_search_term",
-            },
-        )
-
-        try:
-            self.search = SearchClientForTesting(self._db)
-        except Exception as e:
-            self.search = None
-            logging.error(
-                "Unable to set up opensearch index, search tests will be skipped.",
-                exc_info=e,
-            )
-
-    def setup_index(self, new_index):
-        "Create an index and register it to be destroyed during teardown."
-        self.search.setup_index(new_index=new_index)
-        self.indexes.append(new_index)
-
-    def teardown_method(self):
-        if self.search:
-            # Delete the works_index, which is almost always created.
-            if self.search.works_index:
-                self.search.indices.delete(self.search.works_index, ignore=[404])
-            # Delete any other indexes created over the course of the test.
-            for index in self.indexes:
-                self.search.indices.delete(index, ignore=[404])
-            ExternalSearchIndex.reset()
-        super().teardown_method()
-
-    def default_work(self, *args, **kwargs):
-        """Convenience method to create a work with a license pool
-        in the default collection.
-        """
-        work = self._work(
-            *args, with_license_pool=True, collection=self._default_collection, **kwargs
-        )
-        work.set_presentation_ready()
-        return work
-
-
 class MockCoverageProvider:
     """Mixin class for mock CoverageProviders that defines common constants."""
