+++ tests/api/admin/controller/test_collection_self_tests.py
@@ -1,9 +1,10 @@
 from flask_babel import lazy_gettext as _
 
 from api.admin.problem_details import *
-from api.axis import Axis360API, MockAxis360API
+from api.axis import Axis360API
 from core.opds_import import OPDSImportMonitor
 from core.selftest import HasSelfTests
+from tests.api.mockapi.axis import MockAxis360API
 from tests.fixtures.api_admin import SettingsControllerFixture
