+++ tests/api/test_overdrive.py
@@ -15,7 +15,6 @@ from api.circulation import CirculationAPI, FulfillmentInfo, HoldInfo, LoanInfo
 from api.circulation_exceptions import *
 from api.config import Configuration
 from api.overdrive import (
-    MockOverdriveAPI,
     NewTitlesOverdriveCollectionMonitor,
     OverdriveAPI,
     OverdriveCirculationMonitor,
@@ -41,6 +40,7 @@ from core.model import (
 from core.overdrive import OverdriveConfiguration
 from core.testing import DummyHTTPClient, MockRequestsResponse
 from core.util.datetime_helpers import datetime_utc, utc_now
+from tests.api.mockapi.overdrive import MockOverdriveAPI
 
 from ..fixtures.api_overdrive_files import OverdriveAPIFilesFixture
 from ..fixtures.database import DatabaseTransactionFixture
